worker/requirements.txt
python-dotenv
requests
paramiko
autoflake
boto3
isort
black
selenium
